encapsulate external scripts and auto-call scroll event handler
