+++ js/scroller/content/game/blocks.js
@@ -1,5 +1,3 @@
-import * as shapes from './shapes.js';
-
 class Blocks {
   updateView () {
     this.target.innerHTML = `
@@ -166,9 +164,15 @@ class Blocks {
     this.colors = options.colors || [ '#f04', '#0f8', '#08f', '#80f' ];
     this.board = this.initBoard();
     this.pieces = [];
-    this.shapes = [];
+    this.shapes = [
+      ShapeL,
+      ShapeT,
+      ShapeI,
+      ShapeZ,
+      ShapeS,
+      ShapeBox
+    ];
     this.id = Math.random().toString(16).slice(2);
-    for (const shape in shapes) this.shapes.push(shapes[shape]);
     
     this.updateView();
     this.stylize();
@@ -177,4 +181,154 @@ class Blocks {
   }
 }
 
-export default Blocks;
+class Piece {
+  isAtBottom () {
+    return this.y + this.height - 1 === this.game.size;
+  }
+  
+  pieceIsBelow () {
+    const hasCollisions = this.shape.reduce((result, row, y) => {
+      if (result) return result;
+      row.forEach((cell, x) => {
+        const cellHasValue = !!cell;
+        const nextRowInShapeHasValue = !!this.shape[y + 1] ? !!this.shape[y + 1][x] : false;
+        const nextRowOnBoardHasValue = !!this.game.board[this.y + y][this.x + x];
+        if (cellHasValue && nextRowOnBoardHasValue && !nextRowInShapeHasValue) result = true;
+      });
+      return result;
+    }, false);
+    
+    return hasCollisions;
+  }
+  
+  canContinue (specificScenario) {
+    const _this = this;
+    const enders = ['isAtBottom', 'pieceIsBelow'];
+    if (!specificScenario) {
+      const len = enders.length;
+      let i = 0;
+      let ender;
+      
+      for (; (ender = enders[i]) && i < len; ++i) {
+        if (this[ender]()) return false;
+      }
+      
+      return true;
+    } else {
+      const scenario = this[specificScenario];
+      return (scenario && !scenario.call(this)) || false;
+    }
+  }
+  
+  updatePlacement () {
+    if (!this.canContinue()) {
+      this.game.checkForCompleteRow()
+      this.game.active = false;
+    } else {
+      this.y - 1 >= 0 && this.game.board[this.y - 1].splice(this.x, this.width, ...Array(this.width).fill(0));
+      this.shape.forEach((row, y) => {
+        const values = row.map((cell, x) => {
+          const thisShapeHasValue = !!cell;
+          const clearCell = y !== this.shape.length - 1 && !cell;
+          const boardCellValue = this.game.board[this.y + y][this.x + x];
+          return thisShapeHasValue ? this.color : clearCell ? 0 : boardCellValue || 0;
+        });
+        this.game.board[this.y + y].splice(this.x, row.length, ...values);
+      });
+      
+      this.game.updateView();
+      
+      ++this.y;
+      
+      this.y === 1 && !this.canContinue() && this.game.gameOver();
+    }
+  }
+  
+  initOnBoard () {
+    this.width = this.shape[0].length;
+    this.height = this.shape.length;
+    this.x = Math.floor(Math.random() * (this.game.size - this.width));
+    this.y = 0;
+    
+    this.updatePlacement();
+  }
+  
+  constructor (game) {
+    const colors = game.colors;
+      
+    this.game = game;
+    this.element = document.createElement('div');
+    this.color = colors[Math.floor(Math.random() * colors.length)];
+  }
+}
+
+class ShapeL extends Piece {
+  constructor (game) {
+    super(game);
+    this.shape = [
+      [ 0, 1 ],
+      [ 0, 1 ],
+      [ 1, 1 ],
+    ];
+    this.initOnBoard();
+  }
+}
+
+class ShapeT extends Piece {
+  constructor (game) {
+    super(game);
+    this.shape = [
+      [ 1, 1, 1],
+      [ 0, 1, 0 ]
+    ];
+    this.initOnBoard();
+  }
+}
+
+class ShapeZ extends Piece {
+  constructor (game) {
+    super(game);
+    this.shape = [
+      [ 1, 1, 0],
+      [ 0, 1, 1 ]
+    ];
+    this.initOnBoard();
+  }
+}
+
+class ShapeS extends Piece {
+  constructor (game) {
+    super(game);
+    this.shape = [
+      [ 0, 1, 1],
+      [ 1, 1, 0 ]
+    ];
+    this.initOnBoard();
+  }
+}
+
+class ShapeI extends Piece {
+  constructor (game) {
+    super(game);
+    this.shape = [
+      [ 1 ],
+      [ 1 ],
+      [ 1 ],
+      [ 1 ]
+    ];
+    this.initOnBoard();
+  }
+}
+
+class ShapeBox extends Piece {
+  constructor (game) {
+    super(game);
+    this.shape = [
+      [ 1, 1],
+      [ 1, 1 ]
+    ];
+    this.initOnBoard();
+  }
+}
+
+win.Blocks = Blocks;
